Move CMIP6 code into plugins/cmip6; update models/records shims to import from plugin; keep public imports stable
