+++ src/piddiplatsch/plugins/__init__.py
@@ -1,0 +1,1 @@
+# Plugin package namespace for built-in and external plugins.
