+++ src/piddiplatsch/plugins/cmip6/__init__.py
@@ -1,0 +1,1 @@
+# Built-in CMIP6 plugin package.
